requirements.txt
mistralai
openai
google-generativeai
anthropic
python-dotenv
pillow
pdf2image
PyMuPDF
requests
markdown
